Change the type of the line number from unsigned long to unsigned since the tools won't run on a 16 bit platform anyway. And if so, line numbers limited to 65535 are probably the least problem.
git-svn-id: svn://svn.cc65.org/cc65/trunk@5167 b7a2c559-68d2-44c3-8de9-860c34a00d81

+++ src/common/filepos.h
@@ -47,9 +47,9 @@
 /* Type of a file position */
 typedef struct FilePos FilePos;
 struct FilePos {
-    unsigned long   Line;               /* Line */
-    unsigned        Col;                /* Column */
-    unsigned        Name;               /* File */
+    unsigned    Line;               /* Line */
+    unsigned    Col;                /* Column */
+    unsigned    Name;               /* File */
 };
 
 /* Initializer for a FilePos */
